// client/src/components/rac-editor/@canvas/lib/factory/elements/element.strategy.ts
import type {Canvas as FabricCanvas} from 'fabric';
import type {CanvasGroup, CanvasObject} from '@/components/rac-editor/@canvas/lib/canvas.ts';

export type ElementStrategyKey =
  | 'line'
  | 'arrow'
  | 'distance'
  | 'wall'
  | 'streetStraight'
  | 'streetCorner'
  | 'dirtRoad'
  | 'square'
  | 'triangle'
  | 'circle'
  | 'water'
  | 'door'
  | 'stairs'
  | 'fossa'
  | 'tree'
  | 'text';

export interface ElementStrategy<
  TGroup extends CanvasGroup = CanvasGroup,
  TObject extends CanvasObject = CanvasObject,
  TOptions = undefined
> {
  create(canvas: FabricCanvas, options?: TOptions): TGroup | TObject;
}